Fix population in study definition
Add parentheses around logic that contains OR. Currently are no parentheses around `sex = "M" OR sex = "F"` and the `AND` operator naturally takes precedence so the condition is not interpreted correctly. 

For clarity I'm also adding this to the age condition `(age >= 2 AND age <=120)` although this is technically the same.